feat (alias dev): config de aliases para ./src y ts compilador y,  separación de controller con ruta
